Lambda/llm_lambda_test/requirements/layer1.txt
boto3
structlog
pyyaml
fastapi
uvicorn
pydantic==1.10.11
PyPDF2
python-docx
textract==1.6.3